Move x-to-x files to a bridge folder
