+++ lib/bridge/irc-to-matrix.js
@@ -1,17 +1,17 @@
 "use strict";
 var q = require("q");
 
-var matrixLib = require("./mxlib/matrix");
-var store = require("./store");
+var matrixLib = require("../mxlib/matrix");
+var store = require("../store");
 
-var pool = require("./irclib/client-pool");
-var roomModels = require("./models/rooms");
+var pool = require("../irclib/client-pool");
+var roomModels = require("../models/rooms");
 var IrcRoom = roomModels.IrcRoom;
-var actions = require("./models/actions");
-var requests = require("./models/requests");
+var actions = require("../models/actions");
+var requests = require("../models/requests");
 
-var logging = require("./logging");
-var log = logging.get("bridge");
+var logging = require("../logging");
+var log = logging.get("irc-to-matrix");
 
 /**
  * Called when the AS receives an IRC message event.
