+++ lib/bridge/matrix-to-irc.js
@@ -2,22 +2,22 @@
 
 var q = require("q");
 
-var matrixLib = require("./mxlib/matrix");
-var ircLib = require("./irclib/irc");
-var membershiplists = require("./membershiplists");
-var store = require("./store");
+var matrixLib = require("../mxlib/matrix");
+var ircLib = require("../irclib/irc");
+var membershiplists = require("../membershiplists");
+var store = require("../store");
 
-var roomModels = require("./models/rooms");
+var roomModels = require("../models/rooms");
 var MatrixRoom = roomModels.MatrixRoom;
 var IrcRoom = roomModels.IrcRoom;
-var actions = require("./models/actions");
-var users = require("./models/users");
+var actions = require("../models/actions");
+var users = require("../models/users");
 var MatrixUser = users.MatrixUser;
-var requests = require("./models/requests");
-var toIrcLowerCase = require("./irclib/formatting").toIrcLowerCase;
+var requests = require("../models/requests");
+var toIrcLowerCase = require("../irclib/formatting").toIrcLowerCase;
 
-var logging = require("./logging");
-var log = logging.get("bridge");
+var logging = require("../logging");
+var log = logging.get("matrix-to-irc");
 
 var createMatrixUserForIrcUser = function(ircUser, req) {
     var defer = q.defer();
