+++ lib/irc-appservice.js
@@ -7,8 +7,8 @@ module.exports.serviceName = "matrix-appservice-irc";
 
 var q = require("q");
 
-var matrixToIrc = require("./matrix-to-irc.js");
-var ircToMatrix = require("./irc-to-matrix.js");
+var matrixToIrc = require("./bridge/matrix-to-irc.js");
+var ircToMatrix = require("./bridge/irc-to-matrix.js");
 var IrcServer = require("./irclib/server.js").IrcServer;
 var ircLib = require("./irclib/irc.js");
 var matrixLib = require("./mxlib/matrix");
