+++ lib/membershiplists.js
@@ -3,7 +3,7 @@
 "use strict";
 
 var q = require("q");
-var matrixToIrc = require("./matrix-to-irc.js");
+var matrixToIrc = require("./bridge/matrix-to-irc.js");
 var ircLib = require("./irclib/irc.js");
 var matrixLib = require("./mxlib/matrix");
 var MatrixUser = require("./models/users").MatrixUser;
